+++ CHANGELOG.md
@@ -597,6 +597,7 @@
 - [Engine can now recover from serialization failures][5591]
 - [Use sbt runEngineDistribution][5609]
 - [Update to GraalVM 22.3.1][5602]
+- [Cache library bindings to optimize import/export resolution][5700]
 
 [3227]: https://github.com/enso-org/enso/pull/3227
 [3248]: https://github.com/enso-org/enso/pull/3248
@@ -693,6 +694,7 @@
 [5591]: https://github.com/enso-org/enso/pull/5591
 [5609]: https://github.com/enso-org/enso/pull/5609
 [5602]: https://github.com/enso-org/enso/pull/5602
+[5700]: https://github.com/enso-org/enso/pull/5700
 
 # Enso 2.0.0-alpha.18 (2021-10-12)